bins/regions implementation for multibamsummary, npz output. Implement RegionBlocks to balance number of temp files versus bps to process.
